test: [P4PU-1309] fixture refactoring
